+++ CHANGES.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - 2025-08-18
 
+### Added
+
+- `resources.save_pydantic` methods.
+- `context` argument to `resources.expect_pydantic` and `resources.save_pydantic` methods (#11).
+
 ### Changed
 
 - Generated JSON now converts negative zero (`-0.0`) to non-negative zero (`0.0`) by default. The underlying `allow_negative_zero` takes a new argument `prepare_for_json_encode` to change this, but the save and expect methods don't make use of it (#10).